Check index distribution on batches

Report the statistics for each subset and variable compared to the original.

Usage

check_index(pheno, index, omit = NULL)

Arguments

pheno

Data.frame with the sample information.

index

A list of indices indicating which samples go to which subset.

omit

Name of the columns of the pheno that will be omitted.

Value

A matrix with the differences with the original data.

Details

The closer the values are to 0, the less difference is with the original distribution, so it is a better randomization.

See also

Examples

index <- create_subset(50, 24)
metadata <- expand.grid(height = seq(60, 80, 5), weight = seq(100, 300, 50),
                        sex = c("Male","Female"))
check_index(metadata, index)
#>         subgroups
#>            SubSet1   SubSet2    SubSet3
#>   height 2.1404262 0.2985830 0.08758654
#>   weight 2.8129719 3.2208906 4.94078438
#>   sex    0.2672749 0.2672749 0.19631710