Create index of subsets of a data

Index of the samples grouped by batches.

Usage

create_subset(size_data, size_subset = NULL, n = NULL, name = "SubSet")

Arguments

size_data

A numeric value of the amount of samples to distribute.

size_subset

A numeric value with the amount of samples per batch.

n

A numeric value with the number of batches.

name

A character used to name the subsets, either a single one or a vector the same size as n.

Value

A random list of indices of the samples.

See also

Examples

index <- create_subset(100, 50, 2)