Design a batch experiment

Given some samples it distribute them in several batches, trying to have equal number of samples per batch. It can handle both numeric and categorical data.

design(pheno, size_subset, omit = NULL, iterations = 500, name = "SubSet")

Arguments

pheno

Data.frame with the sample information.

size_subset

Numeric value of the number of sample per batch.

omit

Name of the columns of the pheno that will be omitted.

iterations

Numeric value of iterations that will be performed.

name

A character used to name the subsets, either a single one or a vector the same size as n.

Value

The indices of which samples go with which batch.

See also

Examples

data(survey, package = "MASS")
index <- design(survey[, c("Sex", "Smoke", "Age")], size_subset = 50,
                iterations = 50)
index
#> $SubSet1
#>  [1]   4   8   9  19  20  22  26  27  28  41  42  58  61  62  72  77  79  85  87
#> [20]  91  92 105 107 109 115 118 126 127 132 133 138 141 142 143 150 154 164 167
#> [39] 170 174 183 190 197 200 212 216 219 222
#> 
#> $SubSet2
#>  [1]   2  10  12  13  16  17  25  29  32  35  37  40  45  48  53  59  60  63  75
#> [20]  76  81  83  84  89  97 102 113 117 122 124 125 137 146 151 153 155 158 161
#> [39] 165 168 181 191 208 209 217 234 235 236
#> 
#> $SubSet3
#>  [1]   3  18  21  23  34  44  47  52  54  56  70  80  93  99 103 104 106 111 114
#> [20] 119 121 123 134 145 157 160 171 175 177 180 184 186 187 188 189 194 198 201
#> [39] 205 206 207 213 215 226 231 233 237
#> 
#> $SubSet4
#>  [1]   1   6  31  36  49  55  64  65  66  67  71  73  82  88  90  94  95  98 100
#> [20] 101 108 112 130 131 135 136 140 147 148 152 159 166 169 176 192 193 195 199
#> [39] 202 218 220 221 223 225 227 229 230
#> 
#> $SubSet5
#>  [1]   5   7  11  14  15  24  30  33  38  39  43  46  50  51  57  68  69  74  78
#> [20]  86  96 110 116 120 128 129 139 144 149 156 162 163 172 173 178 179 182 185
#> [39] 196 203 204 210 211 214 224 228 232
#>