Evaluates the mean of the numeric values

Looks for the standard deviation of the numeric values

Usage

evaluate_sd(i, pheno)

Arguments

i

List of indices

pheno

Data.frame with the samples

Value

A matrix with the standard deviation value for each column for each subset

See also

Examples

data(survey, package = "MASS")
index <- design(survey[, c("Sex", "Smoke", "Age")], size_subset = 50,
                iterations = 10)
#> Warning: There might be some problems with the data use check_data().
# Note that categorical columns will be omitted:
evaluate_sd(index, survey[, c("Sex", "Smoke", "Age")])
#>      Age 
#> 1.647705