Follow up experiments

If an experiment was carried out with some samples and you want to continue with some other samples later on.

Usage

follow_up(
  original,
  follow_up,
  size_subset,
  omit = NULL,
  old_new = "batch",
  iterations = 500
)

Arguments

original

A data.frame with the information of the samples used originally.

follow_up

A data.frame with the information of the new samples.

size_subset

Numeric value of the number of sample per batch.

omit

Name of the columns of the pheno that will be omitted.

old_new

Name of the column where the batch status will be stored. If it matches the name of a column in original it will be used to find previous batches.

iterations

Numeric value of iterations that will be performed.

Value

A data.frame with the common columns of data, a new column old_new, and a batch column filled with the new batches needed.

See also

Examples

data(survey, package = "MASS")
survey1 <- survey[1:118, ]
survey2 <- survey[119:nrow(survey), ]
folu <- follow_up(survey1, survey2, size_subset = 50, iterations = 10)
#> Warning: There are some problems with the data.
#> Warning: There are some problems with the new samples and the batches.
#> Warning: There are some problems with the new data.
#> Warning: There are some problems with the old data.
#> Warning: Column batch is already present. Did you meant this?