Distribute the sample on the plate

This function assumes that to process the batch the samples are distributes in a plate. Sometimes you know in advance the

spatial(
  index,
  pheno,
  omit = NULL,
  remove_positions = NULL,
  rows = LETTERS[1:5],
  columns = 1:10,
  iterations = 500
)

Arguments

index

A list with the samples on each subgroup, as provided from design() or replicates().

pheno

Data.frame with the sample information.

omit

Name of the columns of the pheno that will be omitted.

remove_positions

Character, name of positions.

rows

Character, name of the rows to be used.

columns

Character, name of the rows to be used.

iterations

Numeric value of iterations that will be performed.

Value

The indices of which samples go with which batch.

Examples

data(survey, package = "MASS")
index <- design(survey[, c("Sex", "Smoke", "Age")], size_subset = 50,
                iterations = 25)
index2 <- spatial(index, survey[, c("Sex", "Smoke", "Age")], iterations = 25)
head(index2)
#> $A1
#> [1]  54 163   4  43 188
#> 
#> $B1
#> [1] 171 172  30  68 149
#> 
#> $C1
#> [1]  25  55 107  70 148
#> 
#> $D1
#> [1]  76 152 192 210  45
#> 
#> $E1
#> [1]  59 225  81 211  52
#> 
#> $A2
#> [1] 197 184  29  93 174
#>