Design a batch experiment

Given some samples it distribute them in several batches, trying to have equal number of samples per batch. It can handle both numeric and categorical data.

Usage

design(pheno, size_subset, omit = NULL, iterations = 500, name = "SubSet")

Arguments

pheno

Data.frame with the sample information.

size_subset

Numeric value of the number of sample per batch.

omit

Name of the columns of the pheno that will be omitted.

iterations

Numeric value of iterations that will be performed.

name

A character used to name the subsets, either a single one or a vector the same size as n.

Value

The indices of which samples go with which batch.

See also

Examples

data(survey, package = "MASS")
index <- design(survey[, c("Sex", "Smoke", "Age")], size_subset = 50,
                iterations = 10)
#> Warning: There might be some problems with the data use check_data().
index
#> $SubSet1
#>  [1]   5   8  10  11  12  19  22  25  30  41  48  49  50  54  63  65  66  70  72
#> [20]  96  98 104 105 107 115 117 118 126 131 132 137 143 148 157 162 164 170 181
#> [39] 196 200 211 222 228 229 230 231 236 237
#> 
#> $SubSet2
#>  [1]   2   4   7  15  16  17  18  27  28  29  31  33  35  40  42  61  62  68  76
#> [20]  78  82  86  88  97 100 109 119 120 124 163 165 168 178 182 183 189 192 193
#> [39] 195 201 203 206 207 213 221 225 227 232
#> 
#> $SubSet3
#>  [1]   9  26  32  37  39  43  44  45  52  55  56  59  60  67  77  79  81  85  90
#> [20]  92 110 113 116 123 127 129 130 134 135 140 141 145 160 161 169 171 174 175
#> [39] 188 190 210 214 217 218 220 223 224
#> 
#> $SubSet4
#>  [1]   1   3  20  21  23  34  38  46  47  51  58  64  69  80  83  84  91  94  99
#> [20] 103 111 112 114 125 136 139 142 144 146 147 152 154 166 167 172 173 177 180
#> [39] 185 186 199 202 204 205 208 219 233
#> 
#> $SubSet5
#>  [1]   6  13  14  24  36  53  57  71  73  74  75  87  89  93  95 101 102 106 108
#> [20] 121 122 128 133 138 149 150 151 153 155 156 158 159 176 179 184 187 191 194
#> [39] 197 198 209 212 215 216 226 234 235
#>