Design a batch experiment

Given some samples it distribute them in several batches, trying to have equal number of samples per batch. It can handle both numeric and categorical data.

Usage

design(pheno, size_subset, omit = NULL, iterations = 500, name = "SubSet")

Arguments

pheno

Data.frame with the sample information.

size_subset

Numeric value of the number of sample per batch.

omit

Name of the columns of the pheno that will be omitted.

iterations

Numeric value of iterations that will be performed.

name

A character used to name the subsets, either a single one or a vector the same size as n.

Value

The indices of which samples go with which batch.

See also

Examples

data(survey, package = "MASS")
index <- design(survey[, c("Sex", "Smoke", "Age")], size_subset = 50,
                iterations = 50)
#> Warning: There might be some problems with the data use check_data().
index
#> $SubSet1
#>  [1]   1   4   5   8   9  14  16  25  26  28  29  37  41  49  51  57  65  67  70
#> [20]  81  83  88  96  97  98 100 103 115 116 118 119 128 138 140 151 152 166 170
#> [39] 172 175 180 187 195 199 207 208 224 230
#> 
#> $SubSet2
#>  [1]   6  34  35  39  40  47  56  58  60  61  63  66  72  73  80  91  94 108 110
#> [20] 111 112 114 117 120 125 134 139 144 146 147 149 154 164 179 182 190 196 201
#> [39] 203 213 214 218 223 226 228 231 232 237
#> 
#> $SubSet3
#>  [1]   2   7  23  24  32  33  42  45  74  85  86  87  89  95 104 107 109 113 122
#> [20] 127 133 135 136 137 142 145 150 157 161 162 165 167 169 171 173 174 181 184
#> [39] 188 191 193 198 206 210 216 219 227
#> 
#> $SubSet4
#>  [1]  10  12  15  17  18  19  22  27  30  31  36  48  52  53  54  55  59  64  68
#> [20]  71  78  79  82  92 106 123 131 141 153 155 160 163 177 186 189 192 194 197
#> [39] 200 209 211 212 215 225 233 234 235
#> 
#> $SubSet5
#>  [1]   3  11  13  20  21  38  43  44  46  50  62  69  75  76  77  84  90  93  99
#> [20] 101 102 105 121 124 126 129 130 132 143 148 156 158 159 168 176 178 183 185
#> [39] 202 204 205 217 220 221 222 229 236
#>