Evaluates a data.frame

Measures several indicators per group

evaluate_index(i, pheno)

Arguments

i
Index

pheno
Data.frame with information about the samples

Value

An array of three dimensions with the mean, standard deviation (sd()), and median absolute deviation (mad()) of the numeric variables, the entropy of the categorical and the number of NA by each subgroup.

See also

Examples

data(survey, package = "MASS")
index <- create_subset(nrow(survey), 50, 5)
ev_index <- evaluate_index(index, survey[, c("Sex", "Smoke")])
ev_index["entropy", , ]

#>          subgroups
#> variables   SubSet1   SubSet2   SubSet3   SubSet4   SubSet5
#>     Sex   0.9987473 0.9987473 0.9918208 0.9986360 0.9996734
#>     Smoke 0.6012896 0.6280066 0.3569229 0.4062612 0.5352755