Distribute the sample on the plate

This function assumes that to process the batch the samples are distributed in a plate with a grid scheme.

Usage

spatial(
  index,
  pheno,
  omit = NULL,
  remove_positions = NULL,
  rows = LETTERS[1:5],
  columns = 1:10,
  iterations = 500
)

Arguments

index

A list with the samples on each subgroup, as provided from design() or replicates().

pheno

Data.frame with the sample information.

omit

Name of the columns of the pheno that will be omitted.

remove_positions

Character, name of positions to be avoided in the grid.

rows

Character, name of the rows to be used.

columns

Character, name of the rows to be used.

iterations

Numeric value of iterations that will be performed.

Value

The indices of which samples go with which batch.

Examples

data(survey, package = "MASS")
index <- design(survey[, c("Sex", "Smoke", "Age")], size_subset = 50,
                iterations = 10)
#> Warning: There might be some problems with the data use check_data().
index2 <- spatial(index, survey[, c("Sex", "Smoke", "Age")], iterations = 10)
head(index2)
#> $A1
#> [1]  86 131  52  76   1
#> 
#> $B1
#> [1]  27 130 124 133 175
#> 
#> $C1
#> [1]  74  21 110 161  99
#> 
#> $D1
#> [1] 162  48 158 108 233
#> 
#> $E1
#> [1] 141 234 222 166 127
#> 
#> $A2
#> [1] 100  49 227  29 125
#>